I think its about time i committed changes.
git-svn-id: https://testsrhoh3aga.myversioncontrol.com/subversion/eahoehaaa2245/trunk@327 e2d0575d-2f82-4250-b8ec-0d5e7f3c5cc0

+++ src/scripts/src/Common/Base.cpp
@@ -173,7 +173,7 @@ bool MoonScriptCreatureAI::GetCanEnterCombat()
 
 void MoonScriptCreatureAI::SetCanEnterCombat(bool pCanEnterCombat)
 {
-	_unit->SetUInt64Value(UNIT_FIELD_FLAGS, ( pCanEnterCombat ) ? 0 : UNIT_FLAG_NOT_ATTACKABLE_9);
+	_unit->SetUInt32Value(UNIT_FIELD_FLAGS, ( pCanEnterCombat ) ? 0 : UNIT_FLAG_NOT_ATTACKABLE_9);
 	_unit->GetAIInterface()->SetAllowedToEnterCombat(pCanEnterCombat);
 }
